new Question on BinarySearch added Search Inseart K in sorted Array

+++ DSA_Questions/practice.cpp
@@ -2,5 +2,5 @@
 using namespace std;
 int main()
 {
-    vector<long long >arr = {1,2,8,10,11,12,19};
+    vector<int >arr = {1,2,8,10,11,12,19};
 }
